organize session context files
