+++ frontend-react/src/utils/SessionStorageTools.ts
